moved RepositoryExtensions to GitExtensions cleanup
