+++ src/GitVersionCore/VersionCalculation/BaseVersionCalculators/TaggedCommitVersionStrategy.cs
@@ -34,10 +34,8 @@ namespace GitVersion.VersionCalculation
                 .SelectMany(commit => { return allTags.Where(t => IsValidTag(t.Item1, commit)); })
                 .Select(t =>
                 {
-                    if (t.Item1.PeeledTarget() is LibGit2Sharp.Commit commit)
-                        return new VersionTaggedCommit((Commit)commit, t.Item2, t.Item1.FriendlyName);
-
-                    return null;
+                    var commit = t.Item1.PeeledTargetCommit();
+                    return commit != null ? new VersionTaggedCommit(commit, t.Item2, t.Item1.FriendlyName) : null;
                 })
                 .Where(versionTaggedCommit => versionTaggedCommit != null)
                 .Select(versionTaggedCommit => CreateBaseVersion(Context, versionTaggedCommit))
@@ -61,7 +59,8 @@ namespace GitVersion.VersionCalculation
 
         protected virtual bool IsValidTag(Tag tag, Commit commit)
         {
-            return tag.PeeledTarget() is LibGit2Sharp.Commit targetCommit && targetCommit == commit;
+            var targetCommit = tag.PeeledTargetCommit();
+            return targetCommit != null && targetCommit == commit;
         }
 
         protected class VersionTaggedCommit
