+++ src/GitVersionCore/VersionCalculation/IncrementStrategyFinder.cs
@@ -2,6 +2,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
+using GitVersion.Extensions;
 using GitVersion.VersionCalculation;
 
 namespace GitVersion
